Added debugging to the mcp server
The server has had a new functionality added in, in place of the old ocr funciton which was performed locally, we have addedd. mistral ocr to the system to provide the highest possible quality of OCR to the documents, when this happened, we introduced errors. that are shwoing in claude desktop when the server is run from there, we have added debugging output here in hopes that we can identity where these errors are coming from and address them in the future.